🐛 - Make ConfigManager platform independent
In this commit, changes were made to ConfigManager.php to make the file paths work on all platforms, not just Unix. Previously, it was only fetching file paths for a Unix system and would have failed on Windows.

Now, it will make use of the getenv function to check for environment variables like HOME or USERPROFILE and use it for file fetching. This will make sure the file paths are working on both Unix and Windows systems. With this change, ConfigManager.php has now become platform independent. This was a necessary change as we were aiming for this project to be cross platform.